+++ changelog.md
@@ -1,6 +1,8 @@
 ## [0.2] - 2023-07-31
 
-- fix "Prod" type
+- fix "Prod" type, add more functions
+- WithT: rename from TransformT, add more instances & functions
+- refRestore etc.: more general because no masking of async exceptions
 
 ## [0.1] - 2022-09-12